Union tablas oneToMany

+++ src/main/java/es/franchise/service/ShopServiceImp.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import es.franchise.dao.IPictureDAO;
 import es.franchise.dao.IShopDAO;
 import es.franchise.dto.Picture;
 import es.franchise.dto.Shop;
@@ -35,12 +36,48 @@ public class ShopServiceImp implements IShopService {
 	public Shop updateShop(Shop shop) {
 		return iShopDao.save(shop);
 	}
-
+	
 	@Override
-	public void deleteShop(int id) {
-		iShopDao.deleteById(id);
-
+	public void deleteShop(int id) {				
+			iShopDao.deleteById(id);
+			
 	}
+	
+	@Override
+	public void deleteShopEmpty(int id) {
+		// SI EL ID DE LA TIENDA ESTA EN LA TABLA PICTURES NO SE PODRA BORRAR
+		
+		Boolean existShop=true,
+				existArtist=true;		
+		
+		
+		PictureServiceImp pictureServiceImp = null;
+		ArtistServiceImp artistServiceImp = null;
+		
+		Shop shop = shopXID(id) ;
+			
+		existShop = pictureServiceImp.pictureXShop(shop);
+		existArtist = artistServiceImp.artistXShop(shop);
+		
+		if(existShop==false && existArtist==false) {
+			
+			iShopDao.deleteById(id);
+			
+		}
+	}
+	
+	@Override
+	public void deleteAllShop() {
+		
+			PictureServiceImp pictureServiceImp = null;
+			ArtistServiceImp artistServiceImp = null;
+		
+			pictureServiceImp.deleteAllPicture();
+			artistServiceImp.deleteAllArtist();
+			
+			iShopDao.deleteAll();			
+		}
+
